+++ topic5/src/main/java/topic5/cartService/Item.java
@@ -2,10 +2,10 @@ package topic5.cartService;
 
 public class Item {
 	private String name;
-	private double price;
+	private int price;
 	private int id;
 	
-	public Item(String name, double price) {
+	public Item(String name, int price) {
 		this.name = name;
 		this.price = price;
 	}
@@ -18,11 +18,11 @@ public class Item {
 		this.name = name;
 	}
 
-	public double getPrice() {
+	public int getPrice() {
 		return price;
 	}
 
-	public void setPrice(double price) {
+	public void setPrice(int price) {
 		this.price = price;
 	}
